data-collection/python/requirements.txt
mysql-connector-python==8.0.32
paho-mqtt==1.6.1
protobuf==3.20.3
